pyproject.toml
[project]
name = "ai-slide-generator"
version = "0.1.0"
description = "AI Slide Generator using OpenAI Agents SDK"
authors = [
    {name = "User",email = "user@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.77.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "openai-agents (>=0.0.14,<0.0.15)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)"
]

[tool.poetry.group.dev.dependencies]
openai = ">=1.77.0,<2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
